% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customs_encounter.R
\docType{data}
\name{cbp_resp}
\alias{cbp_resp}
\title{CBP Response Data}
\format{
A data frame with columns:
\describe{
\item{fiscal_year}{Fiscal year of the encounters.}
\item{demographic}{Demographic group of the encounters (e.g., "Unaccompanied Children").}
\item{encounter_count}{Number of encounters recorded.}
}
}
\source{
CBP public data
}
\usage{
cbp_resp
}
\description{
This dataset contains information about border encounters, categorized by fiscal year, demographics, and other attributes.
}
\keyword{datasets}
