% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customs_encounter.R
\docType{data}
\name{cbp_state}
\alias{cbp_state}
\title{CBP State Data}
\format{
A data frame with columns:
\describe{
\item{state}{State where encounters occurred.}
\item{encounter_count}{Total number of encounters in the state.}
}
}
\source{
CBP public data
}
\usage{
cbp_state
}
\description{
This dataset provides state-level summaries of CBP encounters.
}
\keyword{datasets}
